+++ src/manager/InMemoryHistoryManager.java
@@ -71,15 +71,3 @@ public class InMemoryHistoryManager implements HistoryManager {
         }
     }
 }
-
-class Node<T> {
-    public T data;
-    public Node<T> next;
-    public Node<T> prev;
-
-    public Node(Node<T> prev, T data, Node<T> next) {
-        this.data = data;
-        this.next = next;
-        this.prev = prev;
-    }
-}
