+++ src/manager/Node.java
@@ -1,0 +1,13 @@
+package manager;
+
+class Node<T> {
+    public T data;
+    public Node<T> next;
+    public Node<T> prev;
+
+    public Node(Node<T> prev, T data, Node<T> next) {
+        this.data = data;
+        this.next = next;
+        this.prev = prev;
+    }
+}
